feat: move route to metadata route
